added streaming support to SliceBySliceImageFilter
Utilize the default implementation for GenerateInputRequestedRegion,
which just copies the enlarged output requested region to the input
images. Modified the EnlargeOutputRequestedRegion methods to enlarge
all dimension to the largest except m_Dimensions, so that the filter
can stream in the slice dimension.

The one caveat with these changes are that it's unclear how this
filter behaves when the input and output dimension are different
(the prior code is not bug free with regaurd to this). These changes
are made under the assumption that they are the same dimensions.