ENH: added streaming support to SliceBySliceImageFilter
Utilize the default implementation for GenerateInputRequestedRegion,
which just copies the enlarged output requested region to the input
images. Modified the EnlargeOutputRequestedRegion methods to enlarge
all dimension to the largest except m_Dimensions, so that the filter
can stream in the slice dimension.

The one caveat with these changes are that it's unclear how this
filter behaves when the input and output dimension are different
(the prior code is not bug free with regaurd to this). These changes
are made under the assumption that they are the same dimensions.

DOC: added comments and todos regaurding streaming and parameters

PERF: improved performance by moving looped image allocation to outside loop

The vectors of image internalInputs was being allocated and the image
initialized each loop. This was extracted outside the loop, and
leaving the Image::Allocate methods inside the loop so that it will
still reallocate if needed for in-place filter.

PERF: improved copying of regions and reordered slice dimensions

The dimensions of the slices were reordered to be more intuative and
efficient. They are now simple ordered from hightest to lowest
skipping the slice dimension.

The coping of the regions is now done with image regions iterators,
which is more efficent then the per pixel index and comparisions,
along with being more readable. A currentRegion was also added to
represent the region of the filters input and out region currently
being processed.