improved copying of regions and reordered slice dimensions
The dimensions of the slices were reordered to be more intuative and
efficient. They are now simple ordered from hightest to lowest
skipping the slice dimension.

The coping of the regions is now done with image regions iterators,
which is more efficent then the per pixel index and comparisions,
along with being more readable. A currentRegion was also added to
represent the region of the filters input and out region currently
being processed.