TEMP CHECKIN to show solution to scale-space bug
I have been working on a multi-scale laplacian blob detector, so this
scaling is critical for this to be working correctly. However for a 3D
Laplacian, the normalization does not match any know rational I can
find. Specifically for the 3D case the laplacian is normalized with
sigma^3 not sigma^2. ( Side Note: standard scale space notation of
L(x;t) where t=sigma^2, would indicate that the normalized laplacian =
t(L_xx+L_yy+L_zz) )

From my recent readings there is a "natural coordinates" or
"dimensionless spatial units" which are frequently used for this
analysis. The conclusion I reach from this analysis is that the
correct scaling factor should be sigma^n where n is the order of the
differentiating for the separated convolutions. ( Not the current
constant sigma ).

Additionally the scale space normalization was also scaled by the
pixel spacing which is also incorrect.

Change-Id: I6d796fd7982162e0ce27226c43b92343cd2cff61