WIP: more hacking
WIP: tenative solution to pixel spacing bug in GaussianDerivativeOperator

Corrected pixel spacing, added 1/dx^order normalization outside of
calculation of the kernel, simular to the Recursive Gaussian, not
sure what the best appoarch is here. Need to add explicit tests for
this change, test scale space properties etc.

The path which utilizes the DerivativeOperator has not been looked at.

Change-Id: I7bd2085316b671cde831a4500ef3dbbada0d2765

ENH: terminate operator size when numerically insignificant

Prior the the operator size would continue util tolerance was
reached. This was improved to stop when the contribution was
insignificant numerically. That is when it would not actually add to
the total due to precision.

Change-Id: I933e1d925446d33cda363ba63df22b6d5b78e58d