STYLE: Make MirrorPadImageFilter::DelayBase a conventional parameter
This implementation relies on compilers removal of conditionals based
on boolean template parameters. The value of the if statement is
obvious during compilation time so the dead code is removed during
optimization. The implementation also removes the computation for the
unused distanceFromEdge.

This implementations remove the need for the cumbersome
meta-programing.

Change-Id: I0e2625003ef71135035093615c81cacc7f86ef23