WIP: Refactor logging to include file, linenumber, and object arguments
Change the logging macros to pass the components of the message as
arguments, as opposed to composing them into a string. This will allow
the behavior of the OutputWindow to override how the message is
displayed. This include omitting the details of the filename and the
line number and just displaying the message.